external/bluelet: apply bluelet BIN from libframework to libbluelet
With current build&link rule, if source included, link will get to
app/staging matching the .a file, but for bluelet most internal
devlopers also use lib but not source.

Then if its lib name is libframework.a, linking only find its function in
app/staging, but not verdor/bes...libs, this will cause building error.
Hence we made it a seperate name libbluelet.a

Signed-off-by: haopengxiang <haopengxiang@xiaomi.com>